+++ src/enhancers/catchPixieError.js
@@ -6,7 +6,7 @@ import type {
   TamePixie,
   TamePixieInput,
   WildPixie
-} from '../redux-pixies.js'
+} from '../types.js'
 import { tamePixie } from './tamePixie.js'
 
 type ErrorHandler<P> = (e: any, props: P, onError: OnError) => void
