+++ src/managers/filterPixie.js
@@ -6,7 +6,7 @@ import type {
   TamePixie,
   TamePixieInput,
   WildPixie
-} from '../redux-pixies.js'
+} from '../types.js'
 import { catchify, shallowCompare } from './util.js'
 
 /**
