+++ src/managers/startPixie.js
@@ -2,12 +2,7 @@
 
 import { catchPixieError } from '../enhancers/catchPixieError.js'
 import { reflectPixieOutput } from '../enhancers/reflectPixieOutput.js'
-import type {
-  OnError,
-  OnOutput,
-  PixieInstance,
-  WildPixie
-} from '../redux-pixies.js'
+import type { OnError, OnOutput, PixieInstance, WildPixie } from '../types.js'
 
 function defaultOnError(e: any) {}
 function defaultOnOutput(data: any) {}
