+++ src/managers/util.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { OnError } from '../redux-pixies.js'
+import type { OnError } from '../types.js'
 
 /**
  * If the function throws, send that through `onError`.
